Extract index iterator into a separate file
